+++ ponomarev.viacheslav/S5/funcObj.cpp
@@ -2,15 +2,15 @@
 #include <limits>
 #include <stdexcept>
 
-void ponomarev::Key_summ::operator()(const std::pair< int, std::string > & value)
+void ponomarev::Key_summ::operator()(const std::pair< long long, std::string > & value)
 {
-  bool checkOverflow = (value.first > 0) && (sum_ > (std::numeric_limits< int >::max() - value.first));
+  bool checkOverflow = (value.first > 0) && (sum_ > (std::numeric_limits< long long >::max() - value.first));
   if (checkOverflow)
   {
     throw std::out_of_range("Error: owerflow");
   }
 
-  bool checkUnderflow = (value.first < 0) && (sum_ < (std::numeric_limits< int >::min() - value.first));
+  bool checkUnderflow = (value.first < 0) && (sum_ < (std::numeric_limits< long long >::min() - value.first));
   if (checkUnderflow)
   {
     throw std::out_of_range("Error: underflow");
